+++ src/main/java/io/github/realyusufismail/bentenmod/core/blocks/bententable/IOmnitrixCraftingRecipe.java
@@ -9,4 +9,5 @@ public interface IOmnitrixCraftingRecipe extends Recipe<OmnitrixCrafterContainer
         return RecipeTypeInit.OMNITRIX_CRAFTER_TYPE.get();
     }
 
+    OmnitrixCrafterCraftingBookCategory getCategory();
 }