+++ src/main/java/io/github/realyusufismail/bentenmod/core/init/ItemInit.java
@@ -33,10 +33,9 @@
 package io.github.realyusufismail.bentenmod.core.init;
 
 import io.github.realyusufismail.bentenmod.BenTenMod;
+import io.github.realyusufismail.bentenmod.core.armour.*;
 import io.github.realyusufismail.bentenmod.core.material.CustomArmorMaterial;
 import io.github.realyusufismail.bentenmod.core.material.CustomToolMaterial;
-import io.github.realyusufismail.bentenmod.events.*;
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.*;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
@@ -138,110 +137,114 @@ public class ItemInit {
     // knight armour
     public static final RegistryObject<Item> HELMET =
             ITEMS.register("helmet", () -> new LegendaryArmor(CustomArmorMaterial.ARMOUR,
-                    EquipmentSlot.HEAD, new Item.Properties()));
+                    ArmorItem.Type.HELMET, new Item.Properties()));
 
     public static final RegistryObject<Item> CHESTPLATE =
             ITEMS.register("chestplate", () -> new LegendaryArmor(CustomArmorMaterial.ARMOUR,
-                    EquipmentSlot.CHEST, new Item.Properties()));
+                    ArmorItem.Type.CHESTPLATE, new Item.Properties()));
 
     public static final RegistryObject<Item> LEGGINGS =
             ITEMS.register("leggings", () -> new LegendaryArmor(CustomArmorMaterial.ARMOUR,
-                    EquipmentSlot.LEGS, new Item.Properties()));
+                    ArmorItem.Type.LEGGINGS, new Item.Properties()));
 
     public static final RegistryObject<Item> BOOTS =
             ITEMS.register("boots", () -> new LegendaryArmor(CustomArmorMaterial.ARMOUR,
-                    EquipmentSlot.FEET, new Item.Properties()));
+                    ArmorItem.Type.BOOTS, new Item.Properties()));
     // ben ten clothing
     public static final RegistryObject<Item> JACKET =
             ITEMS.register("jacket", () -> new JacketArmorItem(CustomArmorMaterial.JACKET,
-                    EquipmentSlot.CHEST, new Item.Properties()));
+                    ArmorItem.Type.CHESTPLATE, new Item.Properties()));
 
     // Heatblast armour
 
     public static final RegistryObject<Item> HEATBLAST_HELMET = ITEMS.register("heatblast_helmet",
-            () -> new HeatBlastArmor(CustomArmorMaterial.HEATBLAST_ARMOUR, EquipmentSlot.HEAD,
+            () -> new HeatBlastArmor(CustomArmorMaterial.HEATBLAST_ARMOUR, ArmorItem.Type.HELMET,
                     new Item.Properties().fireResistant().durability(600)));
 
-    public static final RegistryObject<Item> HEATBLAST_CHESTPLATE = ITEMS.register(
-            "heatblast_chestplate", () -> new HeatBlastArmor(CustomArmorMaterial.HEATBLAST_ARMOUR,
-                    EquipmentSlot.CHEST, new Item.Properties().fireResistant().durability(600)));
+    public static final RegistryObject<Item> HEATBLAST_CHESTPLATE =
+            ITEMS.register("heatblast_chestplate",
+                    () -> new HeatBlastArmor(CustomArmorMaterial.HEATBLAST_ARMOUR,
+                            ArmorItem.Type.CHESTPLATE,
+                            new Item.Properties().fireResistant().durability(600)));
 
-    public static final RegistryObject<Item> HEATBLAST_LEGGINGS = ITEMS.register(
-            "heatblast_leggings", () -> new HeatBlastArmor(CustomArmorMaterial.HEATBLAST_ARMOUR,
-                    EquipmentSlot.LEGS, new Item.Properties().fireResistant().durability(600)));
+    public static final RegistryObject<Item> HEATBLAST_LEGGINGS =
+            ITEMS.register("heatblast_leggings",
+                    () -> new HeatBlastArmor(CustomArmorMaterial.HEATBLAST_ARMOUR,
+                            ArmorItem.Type.LEGGINGS,
+                            new Item.Properties().fireResistant().durability(600)));
 
     public static final RegistryObject<Item> HEATBLAST_BOOTS = ITEMS.register("heatblast_boots",
-            () -> new HeatBlastArmor(CustomArmorMaterial.HEATBLAST_ARMOUR, EquipmentSlot.FEET,
+            () -> new HeatBlastArmor(CustomArmorMaterial.HEATBLAST_ARMOUR, ArmorItem.Type.BOOTS,
                     new Item.Properties().fireResistant().durability(600)));
 
     // xlr8 armour
     public static final RegistryObject<Item> XLR8_HELMET =
             ITEMS.register("xlr8_helmet", () -> new Xlr8Armor(CustomArmorMaterial.XLR8_ARMOUR,
-                    EquipmentSlot.HEAD, new Item.Properties()));
+                    ArmorItem.Type.HELMET, new Item.Properties()));
 
     public static final RegistryObject<Item> XLR8_CHESTPLATE =
             ITEMS.register("xlr8_chestplate", () -> new Xlr8Armor(CustomArmorMaterial.XLR8_ARMOUR,
-                    EquipmentSlot.CHEST, new Item.Properties()));
+                    ArmorItem.Type.CHESTPLATE, new Item.Properties()));
 
     public static final RegistryObject<Item> XLR8_LEGGINGS =
             ITEMS.register("xlr8_leggings", () -> new Xlr8Armor(CustomArmorMaterial.XLR8_ARMOUR,
-                    EquipmentSlot.LEGS, new Item.Properties()));
+                    ArmorItem.Type.LEGGINGS, new Item.Properties()));
 
     public static final RegistryObject<Item> XLR8_BOOTS =
             ITEMS.register("xlr8_boots", () -> new Xlr8Armor(CustomArmorMaterial.XLR8_ARMOUR,
-                    EquipmentSlot.FEET, new Item.Properties()));
-    // 4 Arms
+                    ArmorItem.Type.BOOTS, new Item.Properties()));
 
+    // 4 Arms
     public static final RegistryObject<Item> FOURARMS_HELMET = ITEMS.register("fourarms_helmet",
-            () -> new FourArmsArmor(CustomArmorMaterial.FOURARMS_ARMOUR, EquipmentSlot.HEAD,
+            () -> new FourArmsArmor(CustomArmorMaterial.FOURARMS_ARMOUR, ArmorItem.Type.HELMET,
                     new Item.Properties()));
 
     public static final RegistryObject<Item> FOURARMS_CHESTPLATE = ITEMS.register(
             "fourarms_chestplate", () -> new FourArmsArmor(CustomArmorMaterial.FOURARMS_ARMOUR,
-                    EquipmentSlot.CHEST, new Item.Properties()));
+                    ArmorItem.Type.CHESTPLATE, new Item.Properties()));
 
     public static final RegistryObject<Item> FOURARMS_LEGGINGS = ITEMS.register("fourarms_leggings",
-            () -> new FourArmsArmor(CustomArmorMaterial.FOURARMS_ARMOUR, EquipmentSlot.LEGS,
+            () -> new FourArmsArmor(CustomArmorMaterial.FOURARMS_ARMOUR, ArmorItem.Type.LEGGINGS,
                     new Item.Properties()));
 
     public static final RegistryObject<Item> FOURARMS_BOOT = ITEMS.register("fourarms_boot",
-            () -> new FourArmsArmor(CustomArmorMaterial.FOURARMS_ARMOUR, EquipmentSlot.FEET,
+            () -> new FourArmsArmor(CustomArmorMaterial.FOURARMS_ARMOUR, ArmorItem.Type.BOOTS,
                     new Item.Properties()));
 
 
     // INFINITUM Armour
     public static final RegistryObject<Item> INFINITUM_HELMET = ITEMS.register("infinitum_helmet",
-            () -> new ArmorItem(CustomArmorMaterial.INFINITUM_ARMOUR, EquipmentSlot.HEAD,
+            () -> new ArmorItem(CustomArmorMaterial.INFINITUM_ARMOUR, ArmorItem.Type.HELMET,
                     new Item.Properties()));
 
     public static final RegistryObject<Item> INFINITUM_CHESTPLATE = ITEMS
         .register("infinitum_chestplate", () -> new ArmorItem(CustomArmorMaterial.INFINITUM_ARMOUR,
-                EquipmentSlot.CHEST, new Item.Properties()));
+                ArmorItem.Type.CHESTPLATE, new Item.Properties()));
 
     public static final RegistryObject<Item> INFINITUM_LEGGINGS = ITEMS
         .register("infinitum_leggings", () -> new ArmorItem(CustomArmorMaterial.INFINITUM_ARMOUR,
-                EquipmentSlot.LEGS, new Item.Properties()));
+                ArmorItem.Type.LEGGINGS, new Item.Properties()));
 
     public static final RegistryObject<Item> INFINITUM_BOOTS = ITEMS.register("infinitum_boots",
-            () -> new ArmorItem(CustomArmorMaterial.INFINITUM_ARMOUR, EquipmentSlot.FEET,
+            () -> new ArmorItem(CustomArmorMaterial.INFINITUM_ARMOUR, ArmorItem.Type.BOOTS,
                     new Item.Properties()));
     // black diamond armour
     public static final RegistryObject<Item> BLACK_DIAMOND_HELMET = ITEMS.register(
             "black_diamond_helmet", () -> new ArmorItem(CustomArmorMaterial.BLACK_DIAMOND_ARMOUR,
-                    EquipmentSlot.HEAD, new Item.Properties()));
+                    ArmorItem.Type.HELMET, new Item.Properties()));
 
     public static final RegistryObject<Item> BLACK_DIAMOND_CHESTPLATE =
             ITEMS.register("black_diamond_chestplate",
                     () -> new ArmorItem(CustomArmorMaterial.BLACK_DIAMOND_ARMOUR,
-                            EquipmentSlot.CHEST, new Item.Properties()));
+                            ArmorItem.Type.CHESTPLATE, new Item.Properties()));
 
     public static final RegistryObject<Item> BLACK_DIAMOND_LEGGINGS = ITEMS.register(
             "black_diamond_leggings", () -> new ArmorItem(CustomArmorMaterial.BLACK_DIAMOND_ARMOUR,
-                    EquipmentSlot.LEGS, new Item.Properties()));
+                    ArmorItem.Type.LEGGINGS, new Item.Properties()));
 
     public static final RegistryObject<Item> BLACK_DIAMOND_BOOTS = ITEMS.register(
             "black_diamond_boots", () -> new ArmorItem(CustomArmorMaterial.BLACK_DIAMOND_ARMOUR,
-                    EquipmentSlot.FEET, new Item.Properties()));
+                    ArmorItem.Type.BOOTS, new Item.Properties()));
     // Food
     public static final RegistryObject<Item> OMNITRIX_APPLE = ITEMS.register("omnitrix_apple",
             () -> new Item(new Item.Properties().food(FoodInit.OMNITRIX_APPLE)));