+++ src/main/java/io/github/realyusufismail/bentenmod/data/tags/ModItemTagsProvider.java
@@ -50,8 +50,8 @@ public class ModItemTagsProvider extends ItemTagsProvider {
     public ModItemTagsProvider(DataGenerator dataGenerator, BlockTagsProvider blockTagProvider,
             CompletableFuture<HolderLookup.Provider> provider,
             ExistingFileHelper existingFileHelper) {
-        super(dataGenerator.getPackOutput(), provider, blockTagProvider, BenTenMod.MOD_ID,
-                existingFileHelper);
+        super(dataGenerator.getPackOutput(), provider, blockTagProvider.contentsGetter(),
+                BenTenMod.MOD_ID, existingFileHelper);
     }
 
     @Override