+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.19.4-1.0.0] - 20/08/2023
+
+### Ported to 1.19.4
+### Fixed issue with a recipe book
+### Fixed issue with omnitrix crafter
+### Smiting table recipies removed till I make my own smiting table
+### Fixed issue with omnitrix crafter
+
 ## [1.19.3-1.0.4] - 25/02/2023
 
 ### Fixed an issue where blocks where unminable